GC.cpp: PyrGc::Collect: static_cast to uint32 before right shifting
The following code consistently crashes without this static cast.
However, I don't know why this fixes it as it is already checked to be positive.
```
b = Pwhite.new(0, 10, inf).asStream;
Array.fill(1810000, {b.next});`
```

+++ lang/LangSource/GC.cpp
@@ -664,7 +664,7 @@ HOT void PyrGC::Collect() {
     if (mNumToScan > 0) {
         // post("->Collect  ns %d  ng %d  s %d\n", mNumToScan, mNumGrey, mScans);
         // DumpInfo();
-        mNumToScan += mNumToScan >> 3;
+        mNumToScan += static_cast<uint32>(mNumToScan) >> 3;
 
         // post("->Collect2  ns %d  ng %d  s %d\n", mNumToScan, mNumGrey, mScans);
         // mCurSet = 0;
